refactor(permissions): move models into the package, type joins with col()
Move the RBAC table models to app/permissions/models.py (importers
repointed) and replace the join ON-clause type: ignore[arg-type] with
sqlmodel col().
