+++ app/models/__init__.py
@@ -1,6 +1,5 @@
 from app.models.email_verification import EmailVerificationToken
 from app.models.llm import LLMConfig
-from app.models.permissions import Role, RoleAssignment, RolePermission
 from app.models.plugin import Plugin, UserPlugin
 from app.models.user import User
 from app.models.whitelist import WhitelistedEmail
@@ -12,7 +11,4 @@ __all__ = [
     "LLMConfig",
     "WhitelistedEmail",
     "EmailVerificationToken",
-    "Role",
-    "RolePermission",
-    "RoleAssignment",
 ]
